requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Login
Flask-Bcrypt
psycopg2-binary
python-dotenv
pytest
requests # Added for potential API calls later
htmx-flask
email-validator # For WTForms email validation
